Add FastNoise2

+++ src/scenes/world_creation.hpp
@@ -31,7 +31,7 @@ class WorldCreation : public Scene
   World m_world{m_game_context};
   std::unique_ptr<Sprite> m_world_sprite = nullptr;
   WorldRepresentation m_world_representation;
-  std::vector<double> m_height_map{};
+  std::vector<float> m_height_map{};
   InputManager& m_input_manager = InputManager::get_instance();
   int m_seed = 110;
